# docs/mobile-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AIN Mobile API Contract",
    "version": "2026-07-24",
    "description": "Flutter-facing contract for the mobile endpoints currently implemented in the Laravel API."
  },
  "servers": [
    {
      "url": "https://api.example.com/api/v1",
      "description": "Production API base URL placeholder"
    },
    {
      "url": "http://localhost:8000/api/v1",
      "description": "Local Laravel API"
    }
  ],
  "security": [
    {
      "sanctumBearer": []
    }
  ],
  "paths": {
...
    "/organizations/{organizationId}/content": {
      "get": {
        "operationId": "listOrganizationContent",
        "tags": [
          "Content"
        ],
        "summary": "List published organization content visible to the current user, optionally scoped to a room.",
        "parameters": [
          {
            "$ref": "#/components/parameters/OrganizationId"
          },
          {
            "name": "roomId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "perPage",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Visible content items.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessEnvelope"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/ContentItemSummary"
                          }
                        },
                        "meta": {
                          "$ref": "#/components/schemas/PaginationMeta"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/ApiError"
          }
        }
      },
      "post": {
        "operationId": "createOrganizationContentLink",
        "tags": [
          "Organization Content"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/OrganizationId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrganizationContentLinkRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Content link created.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ContentItemSummary"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/organizations/{organizationId}/content/{contentId}": {
      "delete": {
        "operationId": "deleteOrganizationContent",
        "tags": [
          "Organization Content"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/OrganizationId"
          },
          {
            "$ref": "#/components/parameters/ContentId"
          }
        ],
        "responses": {
          "200": {
            "description": "Content deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "deleted": {
                              "type": "boolean"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sanctumBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Sanctum"
      }
    },
    "parameters": {
      "OrganizationId": {
        "name": "organizationId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "ContentId": {
        "name": "contentId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
...
    },
    "responses": {
      "ApiError": {
        "description": "Standard API error envelope.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      }
    },
    "schemas": {
      "SuccessEnvelope": {
        "type": "object",
        "required": [
          "data",
          "meta",
          "requestId"
        ],
        "properties": {
          "data": {},
          "meta": {
            "type": "object"
          },
          "requestId": {
            "type": "string"
          }
        }
      },
...
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message",
              "details",
              "catalog",
              "requestId"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "object"
              },
              "catalog": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/ErrorCatalogEntry"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "requestId": {
                "type": "string"
              }
            }
          }
        }
      },
      "ErrorCatalogEntry": {
        "type": "object",
        "required": [
          "category",
          "retryable",
          "messageEn",
          "messageAr"
        ],
        "properties": {
          "category": {
            "type": "string"
          },
          "retryable": {
            "type": "boolean"
          },
          "messageEn": {
            "type": "string"
          },
          "messageAr": {
            "type": "string"
          }
        }
      },
...
      "PaginationMeta": {
        "type": "object",
        "required": [
          "currentPage",
          "lastPage",
          "perPage",
          "total"
        ],
        "properties": {
          "currentPage": {
            "type": "integer",
            "minimum": 1
          },
          "lastPage": {
            "type": "integer",
            "minimum": 1
          },
          "perPage": {
            "type": "integer",
            "minimum": 1
          },
          "total": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
...
      "ContentItemSummary": {
        "type": "object",
        "required": [
          "id",
          "organization_id",
          "room_id",
          "title",
          "type",
          "status",
          "download_allowed",
          "watermark_enabled"
        ],
        "additionalProperties": true,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "organization_id": {
            "type": "string",
            "format": "uuid"
          },
          "room_id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "type": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "download_allowed": {
            "type": "boolean"
          },
          "watermark_enabled": {
            "type": "boolean"
          },
          "file_asset": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContentFileAssetSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "externalUrl": {
            "type": "string",
            "format": "uri",
            "nullable": true
          }
        }
      },
      "ContentFileAssetSummary": {
        "type": "object",
        "required": [
          "id",
          "original_name",
          "mime_type",
          "size_bytes",
          "status"
        ],
        "additionalProperties": true,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "original_name": {
            "type": "string"
          },
          "mime_type": {
            "type": "string"
          },
          "size_bytes": {
            "type": "integer",
            "minimum": 0
          },
          "status": {
            "type": "string"
          }
        }
      },
...
      "CreateOrganizationContentLinkRequest": {
        "type": "object",
        "required": [
          "roomId",
          "title",
          "type",
          "externalUrl"
        ],
        "properties": {
          "roomId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "minLength": 2,
            "maxLength": 200
          },
          "description": {
            "type": "string",
            "maxLength": 5000,
            "nullable": true
          },
          "type": {
            "type": "string",
            "enum": [
              "link"
            ]
          },
          "externalUrl": {
            "type": "string",
            "format": "uri",
            "maxLength": 2048
          },
          "downloadAllowed": {
            "type": "boolean",
            "default": false
          },
          "watermarkEnabled": {
            "type": "boolean",
            "default": true
          },
          "availableFrom": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "availableUntil": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "published",
              "archived"
            ],
            "default": "published"
          }
        }
      },
...
    }
  }
}